requirements.txt
blinker==1.9.0
boto3==1.40.6
botocore==1.40.6
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cognitojwt==1.4.1
colorama==0.4.6
cryptography==45.0.6
ecdsa==0.19.1
Flask==3.1.1
Flask-Cognito==1.21
fpdf==1.7.2
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
lxml==6.0.0
MarkupSafe==3.0.2
packaging==25.0
pillow==11.3.0
pyasn1==0.6.1
pycparser==2.22
PyJWT==2.10.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-jose==3.5.0
requests==2.32.4
rsa==4.9.1
s3transfer==0.13.1
six==1.17.0
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
